chore(githubSpider): change type from Int to Long

+++ githubSpider/src/commonMain/kotlin/io/github/stream29/githubinsight/Entity.kt
@@ -1,7 +1,7 @@
 package io.github.stream29.githubinsight
 
 data class User(
-    val id: Int,
+    val id: Long,
     val login: String,
     val avatarUrl: String,
     val followersList: List<User>,
@@ -18,16 +18,16 @@ data class User(
     val hireable: Boolean? = null,
     val bio: String? = null,
     val twitterUsername: String? = null,
-    val publicRepos: Int,
-    val publicGists: Int,
+    val publicRepos: Long,
+    val publicGists: Long,
     val createdAt: String,
     val updatedAt: String,
-    val diskUsage: Int? = null,
-    val collaborators: Int? = null
+    val diskUsage: Long? = null,
+    val collaborators: Long? = null
 )
 
 data class Repository(
-    val id: Int,
+    val id: Long,
     val name: String,
     val fullName: String,
     val owner: User? = null,
@@ -35,18 +35,18 @@ data class Repository(
     val description: String? = null,
     val fork: Boolean,
     val forks: List<Repository>,
-    val forksCount: Int,
+    val forksCount: Long,
     val contributors: List<User>,
     val commits: List<Commit>,
     val issues: List<Issue>,
     val issueEvents: List<IssueEvent>,
-    val openIssuesCount: Int,
+    val openIssuesCount: Long,
     val languages: List<String>,
     val releases: List<Release>,
     val stargazers: List<User>,
-    val stargazersCount: Int,
-    val watchersCount: Int,
-    val size: Int,
+    val stargazersCount: Long,
+    val watchersCount: Long,
+    val size: Long,
     val topics: List<String>,
     val hasIssues: Boolean,
     val hasProjects: Boolean,
@@ -62,8 +62,8 @@ data class Repository(
     val updatedAt: String,
     val permissions: Permissions,
     val allowForking: Boolean,
-    val subscribersCount: Int,
-    val networkCount: Int,
+    val subscribersCount: Long,
+    val networkCount: Long,
     val license: License,
 )
 
@@ -82,8 +82,8 @@ data class Assets(
     val nodeId: String,
     val uploader: User,
     val state: String,
-    val size: Int,
-    val downloadCount: Int,
+    val size: Long,
+    val downloadCount: Long,
     val createdAt: String,
     val updatedAt: String,
 )
@@ -114,6 +114,6 @@ data class Organization(
     val description: String,
     val email: String,
     val isVerified: Boolean,
-    val publicRepos: Int,
-    val followers: Int,
+    val publicRepos: Long,
+    val followers: Long,
 )
